v0.0.7 update

+++ core/src/main/java/com/ansdoship/carbonizedpixeldungeon/actors/mobs/Elemental.java
@@ -53,39 +53,59 @@ public abstract class Elemental extends Mob {
 	{
 		HP = HT = 60;
 		defenseSkill = 20;
-		
+
 		EXP = 10;
 		maxLvl = 20;
-		
+
 		flying = true;
 	}
-	
+
+	private boolean summonedALly;
+
 	@Override
 	public int damageRoll() {
-		return Random.NormalIntRange( 20, 25 );
+		if (!summonedALly) {
+			return Random.NormalIntRange(20, 25);
+		} else {
+			int regionScale = Math.max(2, (1 + Dungeon.depth/5));
+			return Random.NormalIntRange(5*regionScale, 5 + 5*regionScale);
+		}
 	}
-	
+
 	@Override
 	public int attackSkill( Char target ) {
-		return 25;
+		if (!summonedALly) {
+			return 25;
+		} else {
+			int regionScale = Math.max(2, (1 + Dungeon.depth/5));
+			return 5 + 5*regionScale;
+		}
 	}
-	
+
+	public void setSummonedALly(){
+		summonedALly = true;
+		//sewers are prison are equivalent, otherwise scales as normal (2/2/3/4/5)
+		int regionScale = Math.max(2, (1 + Dungeon.depth/5));
+		defenseSkill = 5*regionScale;
+		HT = 15*regionScale;
+	}
+
 	@Override
 	public int drRoll() {
 		return Random.NormalIntRange(0, 5);
 	}
-	
-	private int rangedCooldown = Random.NormalIntRange( 3, 5 );
-	
+
+	protected int rangedCooldown = Random.NormalIntRange( 3, 5 );
+
 	@Override
 	protected boolean act() {
 		if (state == HUNTING){
 			rangedCooldown--;
 		}
-		
+
 		return super.act();
 	}
-	
+
 	@Override
 	protected boolean canAttack( Char enemy ) {
 		if (rangedCooldown <= 0) {
@@ -94,15 +114,15 @@ public abstract class Elemental extends Mob {
 			return super.canAttack( enemy );
 		}
 	}
-	
+
 	protected boolean doAttack( Char enemy ) {
-		
+
 		if (Dungeon.level.adjacent( pos, enemy.pos ) || rangedCooldown > 0) {
-			
+
 			return super.doAttack( enemy );
-			
+
 		} else {
-			
+
 			if (sprite != null && (sprite.visible || enemy.sprite.visible)) {
 				sprite.zap( enemy.pos );
 				return false;
@@ -112,34 +132,34 @@ public abstract class Elemental extends Mob {
 			}
 		}
 	}
-	
+
 	@Override
 	public int attackProc( Char enemy, int damage ) {
 		damage = super.attackProc( enemy, damage );
 		meleeProc( enemy, damage );
-		
+
 		return damage;
 	}
-	
+
 	private void zap() {
 		spend( 1f );
-		
+
 		if (hit( this, enemy, true )) {
-			
+
 			rangedProc( enemy );
-			
+
 		} else {
 			enemy.sprite.showStatus( CharSprite.NEUTRAL,  enemy.defenseVerb() );
 		}
 
 		rangedCooldown = Random.NormalIntRange( 3, 5 );
 	}
-	
+
 	public void onZapComplete() {
 		zap();
 		next();
 	}
-	
+
 	@Override
 	public void add( Buff buff ) {
 		if (harmfulBuffs.contains( buff.getClass() )) {
@@ -148,42 +168,48 @@ public abstract class Elemental extends Mob {
 			super.add( buff );
 		}
 	}
-	
+
 	protected abstract void meleeProc( Char enemy, int damage );
 	protected abstract void rangedProc( Char enemy );
-	
+
 	protected ArrayList<Class<? extends Buff>> harmfulBuffs = new ArrayList<>();
-	
+
 	private static final String COOLDOWN = "cooldown";
-	
+	private static final String SUMMONED_ALLY = "summoned_ally";
+
 	@Override
 	public void storeInBundle( Bundle bundle ) {
 		super.storeInBundle( bundle );
 		bundle.put( COOLDOWN, rangedCooldown );
+		bundle.put( SUMMONED_ALLY, summonedALly);
 	}
-	
+
 	@Override
 	public void restoreFromBundle( Bundle bundle ) {
 		super.restoreFromBundle( bundle );
 		if (bundle.contains( COOLDOWN )){
 			rangedCooldown = bundle.getInt( COOLDOWN );
 		}
+		summonedALly = bundle.getBoolean( SUMMONED_ALLY );
+		if (summonedALly){
+			setSummonedALly();
+		}
 	}
-	
+
 	public static class FireElemental extends Elemental {
-		
+
 		{
 			spriteClass = ElementalSprite.Fire.class;
-			
+
 			loot = new PotionOfLiquidFlame();
 			lootChance = 1/8f;
-			
+
 			properties.add( Property.FIERY );
-			
+
 			harmfulBuffs.add( com.ansdoship.carbonizedpixeldungeon.actors.buffs.Frost.class );
 			harmfulBuffs.add( Chill.class );
 		}
-		
+
 		@Override
 		protected void meleeProc( Char enemy, int damage ) {
 			if (Random.Int( 2 ) == 0 && !Dungeon.level.water[enemy.pos]) {
@@ -191,7 +217,7 @@ public abstract class Elemental extends Mob {
 				if (enemy.sprite.visible) Splash.at( enemy.sprite.center(), sprite.blood(), 5);
 			}
 		}
-		
+
 		@Override
 		protected void rangedProc( Char enemy ) {
 			if (!Dungeon.level.water[enemy.pos]) {
@@ -200,49 +226,67 @@ public abstract class Elemental extends Mob {
 			if (enemy.sprite.visible) Splash.at( enemy.sprite.center(), sprite.blood(), 5);
 		}
 	}
-	
+
 	//used in wandmaker quest
 	public static class NewbornFireElemental extends FireElemental {
-		
+
 		{
 			spriteClass = ElementalSprite.NewbornFire.class;
-			
+
 			HT = 60;
 			HP = HT/2; //30
-			
+
 			defenseSkill = 12;
-			
+
 			EXP = 7;
-			
+
 			properties.add(Property.MINIBOSS);
+
+			//newborn elementals do not have ranged attacks
+			rangedCooldown = Integer.MAX_VALUE;
 		}
 
 		@Override
 		public void die(Object cause) {
 			super.die(cause);
-			Dungeon.level.drop( new Embers(), pos ).sprite.drop();
+			if (alignment == Alignment.ENEMY) Dungeon.level.drop( new Embers(), pos ).sprite.drop();
 		}
 
 		@Override
 		public boolean reset() {
 			return true;
 		}
-		
+
 	}
-	
+
+	//not a miniboss, fully HP, otherwise a newborn elemental
+	public static class AllyNewBornElemental extends NewbornFireElemental {
+
+		{
+			HP = HT;
+			properties.remove(Property.MINIBOSS);
+		}
+
+		@Override
+		public boolean reset() {
+			return false;
+		}
+
+	}
+
 	public static class FrostElemental extends Elemental {
-		
+
 		{
 			spriteClass = ElementalSprite.Frost.class;
-			
+
 			loot = new PotionOfFrost();
 			lootChance = 1/8f;
-			
+
 			properties.add( Property.ICY );
-			
+
 			harmfulBuffs.add( Burning.class );
 		}
-		
+
 		@Override
 		protected void meleeProc( Char enemy, int damage ) {
 			if (Random.Int( 3 ) == 0 || Dungeon.level.water[enemy.pos]) {
@@ -250,35 +294,35 @@ public abstract class Elemental extends Mob {
 				if (enemy.sprite.visible) Splash.at( enemy.sprite.center(), sprite.blood(), 5);
 			}
 		}
-		
+
 		@Override
 		protected void rangedProc( Char enemy ) {
 			Freezing.freeze( enemy.pos );
 			if (enemy.sprite.visible) Splash.at( enemy.sprite.center(), sprite.blood(), 5);
 		}
 	}
-	
+
 	public static class ShockElemental extends Elemental {
-		
+
 		{
 			spriteClass = ElementalSprite.Shock.class;
-			
+
 			loot = new ScrollOfRecharging();
 			lootChance = 1/4f;
-			
+
 			properties.add( Property.ELECTRIC );
 		}
-		
+
 		@Override
 		protected void meleeProc( Char enemy, int damage ) {
 			ArrayList<Char> affected = new ArrayList<>();
 			ArrayList<Lightning.Arc> arcs = new ArrayList<>();
 			Shocking.arc( this, enemy, 2, affected, arcs );
-			
+
 			if (!Dungeon.level.water[enemy.pos]) {
 				affected.remove( enemy );
 			}
-			
+
 			for (Char ch : affected) {
 				ch.damage( Math.round( damage * 0.4f ), this );
 			}
@@ -293,7 +337,7 @@ public abstract class Elemental extends Mob {
 				Sample.INSTANCE.play(Assets.Sounds.LIGHTNING);
 			}
 		}
-		
+
 		@Override
 		protected void rangedProc( Char enemy ) {
 			Buff.affect( enemy, Blindness.class, Blindness.DURATION/2f );
@@ -302,32 +346,32 @@ public abstract class Elemental extends Mob {
 			}
 		}
 	}
-	
+
 	public static class ChaosElemental extends Elemental {
-		
+
 		{
 			spriteClass = ElementalSprite.Chaos.class;
-			
+
 			loot = new ScrollOfTransmutation();
 			lootChance = 1f;
 		}
-		
+
 		@Override
 		protected void meleeProc( Char enemy, int damage ) {
 			CursedWand.cursedEffect(null, this, enemy);
 		}
-		
+
 		@Override
 		protected void rangedProc( Char enemy ) {
 			CursedWand.cursedEffect(null, this, enemy);
 		}
 	}
-	
+
 	public static Class<? extends Elemental> random(){
 		if (Random.Int( 50 ) == 0){
 			return ChaosElemental.class;
 		}
-		
+
 		float roll = Random.Float();
 		if (roll < 0.4f){
 			return FireElemental.class;
